Test mysql/elastic client must not be called when disabled via searchEngine config prop

+++ clearflask-server/src/test/java/com/smotana/clearflask/testutil/AbstractIT.java
@@ -9,22 +9,26 @@ import com.carrotsearch.randomizedtesting.RandomizedRunner;
 import com.carrotsearch.randomizedtesting.annotations.ThreadLeakScope;
 import com.google.inject.AbstractModule;
 import com.google.inject.Inject;
+import com.google.inject.Module;
 import com.google.inject.util.Modules;
 import com.kik.config.ice.ConfigSystem;
 import com.smotana.clearflask.billing.KillBillClientProvider;
+import com.smotana.clearflask.store.ProjectStore.SearchEngine;
 import com.smotana.clearflask.store.elastic.DefaultElasticSearchProvider;
 import com.smotana.clearflask.store.mysql.DefaultMysqlProvider;
 import com.smotana.clearflask.util.IdUtil;
-import com.smotana.clearflask.web.Application;
 import lombok.extern.slf4j.Slf4j;
 import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
 import org.elasticsearch.client.RequestOptions;
 import org.elasticsearch.client.RestHighLevelClient;
+import org.jooq.DSLContext;
 import org.junit.Before;
 import org.junit.runner.RunWith;
 import org.killbill.billing.client.KillBillHttpClient;
+import org.mockito.Mockito;
 
 import java.net.ConnectException;
+import java.util.Optional;
 
 import static com.carrotsearch.randomizedtesting.annotations.ThreadLeakScope.Scope.NONE;
 import static org.junit.Assert.fail;
@@ -45,19 +49,21 @@ public abstract class AbstractIT extends AbstractTest {
 
         bind(AWSCredentialsProvider.class).toInstance(new AWSStaticCredentialsProvider(new BasicAWSCredentials("", "")));
 
+        Optional<SearchEngine> searchEngineOverrideOpt = Optional.ofNullable(overrideSearchEngine());
+        boolean enableElasticClient = searchEngineOverrideOpt.map(SearchEngine::isWriteElastic).orElse(true);
+        boolean enableMysqlClient = searchEngineOverrideOpt.map(SearchEngine::isWriteMysql).orElse(true);
         install(Modules.override(
-                DefaultElasticSearchProvider.module(),
-                DefaultMysqlProvider.module(),
+                enableElasticClient ? DefaultElasticSearchProvider.module() : disabledElasticClient(),
+                enableMysqlClient ? DefaultMysqlProvider.module() : disabledMysqlClient(),
                 KillBillClientProvider.module()
         ).with(new AbstractModule() {
             @Override
             protected void configure() {
-                install(ConfigSystem.overrideModule(Application.Config.class, om -> {
-                    om.override(om.id().createIndexesOnStartup()).withValue(true);
-                }));
-                install(ConfigSystem.overrideModule(DefaultElasticSearchProvider.Config.class, om -> {
-                    om.override(om.id().serviceEndpoint()).withValue("http://localhost:9200");
-                }));
+                if (enableElasticClient) {
+                    install(ConfigSystem.overrideModule(DefaultElasticSearchProvider.Config.class, om -> {
+                        om.override(om.id().serviceEndpoint()).withValue("http://localhost:9200");
+                    }));
+                }
                 String apiKey = IdUtil.randomAscId();
                 String secretKey = IdUtil.randomId();
                 log.info("KillBill test randomized apiKey {} secretKey {}", apiKey, secretKey);
@@ -70,17 +76,41 @@ public abstract class AbstractIT extends AbstractTest {
                     om.override(om.id().apiSecret()).withValue(secretKey);
                     om.override(om.id().requireTls()).withValue(false);
                 }));
-                String databaseName = "clearflask" + IdUtil.randomId(5);
-                log.info("IT test with mysql database name {}", databaseName);
-                install(ConfigSystem.overrideModule(DefaultMysqlProvider.Config.class, om -> {
-                    om.override(om.id().host()).withValue("localhost");
-                    om.override(om.id().user()).withValue("root");
-                    om.override(om.id().pass()).withValue("killbill");
-                    om.override(om.id().databaseName()).withValue(databaseName);
-                    om.override(om.id().dropDatabase()).withValue(true);
-                }));
+                if (enableMysqlClient) {
+                    String databaseName = "clearflask" + IdUtil.randomId(5);
+                    log.info("IT test with mysql database name {}", databaseName);
+                    install(ConfigSystem.overrideModule(DefaultMysqlProvider.Config.class, om -> {
+                        om.override(om.id().host()).withValue("localhost");
+                        om.override(om.id().user()).withValue("root");
+                        om.override(om.id().pass()).withValue("killbill");
+                        om.override(om.id().databaseName()).withValue(databaseName);
+                        om.override(om.id().dropDatabase()).withValue(true);
+                    }));
+                }
             }
         }));
+    }
+
+    private Module disabledMysqlClient() {
+        return new AbstractModule() {
+            @Override
+            protected void configure() {
+                bind(DSLContext.class).toInstance(Mockito.mock(DSLContext.class, invocation -> {
+                    throw new RuntimeException("Mysql client is disabled, it should never be used");
+                }));
+            }
+        };
+    }
+
+    private Module disabledElasticClient() {
+        return new AbstractModule() {
+            @Override
+            protected void configure() {
+                bind(RestHighLevelClient.class).toInstance(Mockito.mock(RestHighLevelClient.class, invocation -> {
+                    throw new RuntimeException("ElasticSearch client is disabled, it should never be used");
+                }));
+            }
+        };
     }
 
     @Before
